refactor(rendering): :recycle: transfer projector, fov and nearfar from window to camera
